harvest turrets minor rework
