+++ objects/kheAA/kheAA_autoBeamer/fu_harvestTurret.object
@@ -150,7 +150,7 @@
 	"animation": "standingturret.animation",
 
 	"scripts": [
-		"/objects/kheAA/kheAA_autoBeamer/standingturret.lua",
+		"/objects/kheAA/kheAA_autoBeamer/harvestTurret.lua",
 		"/scripts/npcToyObject.lua",
 		"/scripts/stateMachine.lua"
 	],
@@ -171,6 +171,13 @@
 	"maxFireAngle": 1,
 	
 	"fireTime": 1,
+
+	"targetEntityTypes":[
+		"object","plant"
+	],
+	"validTargetTypes":{
+		"farmable":true
+	},
 
 	"npcToy": {
 		"influence": [
